requirements.txt
requests
pygithub
google-generativeai
python-dotenv
